FIX: Fix test
